[ui] Support for Job metrics v2
[INTERNAL_BRANCH=dish/oper-2194-update-designs-for-job]
