+++ js_modules/dagster-ui/packages/ui-core/src/assets/insights/AssetCatalogInsights.module.css
@@ -1,82 +1,8 @@
-.TileContainer {
-  position: relative;
-  width: 100%;
-  height: 100%;
-  border-radius: 2px;
-  cursor: pointer;
-}
-
-.placeholderTile {
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  border-radius: 2px;
-  background-color: var(--color-background-light);
-}
-
-.TileContainer:hover .placeholderTile {
-  box-shadow: inset 0 0 0 1px var(--tile-hover-color);
-}
-
-.tileButton {
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  border-radius: 2px;
-  border: 0;
-  transition: background-color 150ms;
-  background-color: var(--tile-color);
-  cursor: pointer;
-  padding: 0;
-}
-
-.tileButton:hover {
-  background-color: var(--tile-hover-color);
-}
-
-.Tooltip {
-  padding: 4px 8px;
-  border-radius: 4px;
-}
-
-.ActivityChartContainer {
-  padding: 16px 24px;
-  border-radius: 12px;
-  border: 1px solid var(--color-keyline-default);
-}
-
-.ActivityChartRow {
-  display: grid;
-  grid-template-columns: 42px 1fr;
-  gap: 12px;
-}
-
-.ActivityChartBottomLegend {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  align-self: stretch;
-  gap: 6px;
-}
-
-.MainLineCharts {
-  display: grid;
-  grid-template-columns: repeat(2, minmax(0, 1fr));
-  gap: 12px;
-}
-
 .LineCharts {
   display: grid;
   grid-template-columns: repeat(4, minmax(0, 1fr));
   gap: 12px;
   padding-bottom: 12px;
-}
-
-.ActivityChart {
-  display: grid;
-  grid-template-columns: 1fr;
-  gap: 2px;
-  color: var(--color-text-light);
 }
 
 .TopAssetsChartsSection {
@@ -89,8 +15,4 @@
   display: grid;
   grid-template-columns: repeat(3, minmax(0, 1fr));
   gap: 12px;
-}
-
-.Popover {
-  pointer-events: none;
 }
