+++ js_modules/dagster-ui/packages/ui-core/src/insights/css/LineChartWithComparison.module.css
@@ -1,0 +1,40 @@
+.chartContainer {
+  background-color: var(--color-background-default);
+  border-radius: 12px;
+  border: 1px solid var(--color-keyline-default);
+  padding: 16px;
+  width: 100%;
+  height: auto;
+  box-sizing: border-box;
+}
+
+.chartWrapper {
+  position: relative;
+  width: 100%;
+}
+
+.chartGraph {
+  flex-grow: 1;
+  height: 160px;
+  cursor: pointer;
+}
+
+.chartHeader {
+  font-size: 16px;
+  margin-bottom: 12px;
+}
+
+.chartCount {
+  font-size: 24px;
+  font-weight: 500;
+  margin-bottom: 2px;
+  display: flex;
+  align-items: baseline;
+  gap: 4px;
+}
+
+.chartChange {
+  font-size: 14px;
+  color: var(--color-text-lighter);
+  margin-bottom: 16px;
+}
